(#21285) Bump to win32-dir v0.4.3
Previously, running facter on a non-en_US system would fail when calling

    File.join(Dir::COMMON_APPDATA, '...')

because the COMMON_APPDATA constant was UTF-16LE encoded, which File.join
does not accept.

This commit updates the win32-dir gem, which encodes the CSIDL constants
using the default_external encoding.

This commit introduces a new dependency on ffi, which is licensed as LGPLv3.
The ffi gem is also only supported on ruby 1.9 and up.