(#22493) Unable to start puppet agent on non-English Windows
This fixes the inability to install the MSI by patching win32-dir to
allow the default encoding of the machine to be used when no encoding
is determined. Ruby when querying the system SHGetFileInfo returns an
ascii 8bit string and this forces it to use whatever the system encoding
default is set to. This allows puppet to install the agent and run properly.

This does not fix the use of non-en_us ASCII characters in the manifests.

Paired-with: Josh Cooper <josh@puppetlabs.com>