(PUP-1284) Return SID correctly for Authenticated Users
When returning the SID for authenticated users (and a couple of other users),
the SID is returned incorrectly as S-1-5-0 (should be S-1-5-11). This is due to
stripping the SID instead of shortening the SID with GetSIDLength. Certain
characters (ASCII 11) look like whitespace and the way win32-security gem works
in the current version is to strip that off, returning an incorrect SID.

This provides a fix until we upgrade win32-security gem.

Details at https://github.com/djberg96/win32-security/issues/3