(FACT-570) Remove win32-security authenticated users patch
Previously, we were patching win32-security version 0.1.4, because
it called strip on a binary string sid. As a result, the gem would
truncate SIDs ending with whitespace characters, e.g. 'Authenticated
Users' is 'S-1-5-11', and ASCII character 11 is a vertical tab.

Our patch was merged upstream in version 0.2.3[1] and works
correctly in 0.2.5:

    $ cmd /c gem list win32-security
    win32-security (0.2.5)
    $ ruby.exe -r win32/security -e "puts Win32::Security::SID.new('Authenticated Users')"
    S-1-5-11

https://github.com/djberg96/win32-security/commit/e197c389957d11397b5ffce024ee98924b838d7e