(RE-2180) Update to our build of ruby
This replaces the ruby20 build in puppet-win32 ruby with that built by
our jenkins job. Previously, we had the ruby from oneclick, the upstream
maintained of the rubyinstaller. Since we have specific patches we want
to apply, as well as other small changes, we want to have ruby built
from our fork of the ruby installer. These files are pulled from
http://jenkins-release.delivery.puppetlabs.net/job/WindowsRubyBuilds/52/RUBYVER=2.0.0-p481,label=win-build/